should have a working copy of a simple version of the game

+++ Assets/Scripts/Characters/CharacterBase.cs
@@ -90,5 +90,5 @@ public class CharacterBase
     {
         health -= damage;
     }
-    public float GetHealthPercentage() {  return (float)(health / maxHp); }
+    public float GetHealthPercentage() { return ((float)health / (float)maxHp); }
 }
